Refactoring GitRepository class and removing gittle and gitpython references from it
