--- components/org.apache.stratos.python.cartridge.agent/cartridgeagent/cartridgeagent/modules/artifactmgt/git/gitrepository.py
@@ -1,48 +1,0 @@
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-# KIND, either express or implied.  See the License for the
-# specific language governing permissions and limitations
-# under the License.
-
-
-class GitRepository:
-    """
-    Represents a git repository inside a particular instance
-    """
-
-    def __init__(self):
-        self.repo_url = None
-        """ :type : str  """
-        self.local_repo_path = None
-        """ :type : str  """
-        self.cloned = False
-        """ :type : bool  """
-        self.repo = None
-        """ :type : git.repo.base.Repo  """
-        self.gittle_repo = None
-        """ :type : gittle.gittle.Gittle  """
-        self.tenant_id = None
-        """ :type : int  """
-        self.key_based_auth = False
-        """ :type : bool  """
-        self.repo_username = None
-        """ :type : str  """
-        self.repo_password = None
-        """ :type : str  """
-        self.is_multitenant = False
-        """ :type : bool  """
-        self.commit_enabled = False
-        """ :type : bool  """
-        self.scheduled_update_task = None
-        """:type : ScheduledExecutor """
