+++ components/org.apache.stratos.python.cartridge.agent/cartridgeagent/cartridgeagent/modules/artifactmgt/repository.py
@@ -15,7 +15,8 @@
 # specific language governing permissions and limitations
 # under the License.
 
-class RepositoryInformation:
+
+class Repository:
     """
     Holds repository information to be used in artifact management
     """
